hook_funcsを分解し、hook毎にファイルにした。その他リファクタリング
